+++ typings/cli-table2/cli-table2.d.ts
@@ -1,4 +1,0 @@
-declare module "cli-table2" {
-  var p: any;
-  export = p;
-}